Move FullscreenMessage to components dir
